Bulk new features
Starting to work towards live integration with real rotor controls.

Added hard coded (this will change to be user selectable) Serial Port
DCL commands common to a lot of serial port controlled rotors. (not user
selectable yet tho)

Added Moon Track Tool (to main menu) works with the sensor as long as
WSJT-10 is running.

Added Presets, and the ability to change presets (right click on a
preset and you can change it).  Left click attempts to move the rotor to
it.  - Not completed yet!

And a mass of other tweaks and fixes.